requirements/dev.txt
black==19.10b0
pylint==2.4.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
mypy==0.770
pytest==5.4.1
coverage==5.1
